+++ client/cody/src/non-stop/utils.ts
@@ -1,12 +1,11 @@
 export enum CodyTaskState {
     'idle' = 0,
-    'queued' = 1,
+    'waiting' = 1,
     'asking' = 2,
-    'marking' = 3,
-    'ready' = 4,
-    'applying' = 5,
-    'fixed' = 6,
-    'error' = 7,
+    'ready' = 3,
+    'applying' = 4,
+    'fixed' = 5,
+    'error' = 6,
 }
 
 export type CodyTaskList = {
@@ -16,6 +15,7 @@ export type CodyTaskList = {
         description: string
     }
 }
+
 /**
  * Icon for each task state
  */
@@ -23,44 +23,41 @@ export const fixupTaskList: CodyTaskList = {
     [CodyTaskState.idle]: {
         id: 'idle',
         icon: 'clock',
-        description: 'Initial state - all task starts from here',
+        description: 'Initial state',
+    },
+    [CodyTaskState.waiting]: {
+        id: 'waiting',
+        // TODO: Per Figma, this should be a Cody wink
+        icon: 'debug-pause',
+        description: 'The task is waiting to be processed by Cody',
     },
     [CodyTaskState.asking]: {
         id: 'asking',
         icon: 'sync~spin',
-        description: 'In the process of pending for Cody response',
+        description: 'Cody is preparing a response',
     },
     [CodyTaskState.ready]: {
         id: 'ready',
-        icon: 'smiley',
+        icon: 'pencil',
         description: 'Cody has responsed with suggestions and is ready to apply them',
-    },
-    [CodyTaskState.error]: {
-        id: 'error',
-        icon: 'stop',
-        description: 'The task has been completed and returned error',
-    },
-    [CodyTaskState.queued]: {
-        id: 'queue',
-        icon: 'debug-pause',
-        description: 'The task is in the queue to be processed by Cody',
     },
     [CodyTaskState.applying]: {
         id: 'applying',
-        icon: 'sync~spin',
-        description: 'In the process of applying the fixups to the docs',
-    },
-    [CodyTaskState.marking]: {
-        id: 'marking',
-        icon: 'edit',
-        description: 'Marking the fixups as Cody responses',
+        icon: 'pencil',
+        description: 'The fixup is being applied to the document',
     },
     [CodyTaskState.fixed]: {
         id: 'fixed',
         icon: 'pass-filled',
-        description: 'Suggestions from Cody have been applied to the docs successfully',
+        description: 'Suggestions from Cody have been applied or discarded',
+    },
+    [CodyTaskState.error]: {
+        id: 'error',
+        icon: 'stop',
+        description: 'The task failed',
     },
 }
+
 /**
  * Get the last part of the file path after the last slash
  */