+++ client/web/src/repo/blob/own/FileOwnershipPanel.tsx
@@ -7,7 +7,6 @@ import { useQuery } from '@sourcegraph/http-client'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { ErrorAlert, LoadingSpinner } from '@sourcegraph/wildcard'
 
-import { useFeatureFlag } from '../../../featureFlags/useFeatureFlag'
 import { FetchOwnershipResult, FetchOwnershipVariables } from '../../../graphql-operations'
 import { OwnershipAssignPermission } from '../../../rbac/constants'
 
@@ -36,7 +35,6 @@ export const FileOwnershipPanel: React.FunctionComponent<OwnershipPanelProps & T
         },
     })
     const [makeOwnerError, setMakeOwnerError] = React.useState<Error | undefined>(undefined)
-    const [ownPromotionEnabled] = useFeatureFlag('own-promote')
 
     if (loading) {
         return (
@@ -48,18 +46,17 @@ export const FileOwnershipPanel: React.FunctionComponent<OwnershipPanelProps & T
     const canAssignOwners = (data?.currentUser?.permissions?.nodes || []).some(
         permission => permission.displayName === OwnershipAssignPermission
     )
-    const makeOwnerButton =
-        canAssignOwners && ownPromotionEnabled
-            ? (userId: string | undefined) => (
-                  <MakeOwnerButton
-                      onSuccess={refetch}
-                      onError={setMakeOwnerError}
-                      repoId={repoID}
-                      path={filePath}
-                      userId={userId}
-                  />
-              )
-            : undefined
+    const makeOwnerButton = canAssignOwners
+        ? (userId: string | undefined) => (
+              <MakeOwnerButton
+                  onSuccess={refetch}
+                  onError={setMakeOwnerError}
+                  repoId={repoID}
+                  path={filePath}
+                  userId={userId}
+              />
+          )
+        : undefined
 
     if (error) {
         logger.log(error)