+++ client/web/src/site-admin/flags.ts
@@ -5,5 +5,3 @@ export const isPackagesEnabled = (): boolean =>
     window.context?.experimentalFeatures?.rubyPackages === 'enabled' ||
     window.context?.experimentalFeatures?.pythonPackages === 'enabled' ||
     window.context?.experimentalFeatures?.rustPackages === 'enabled'
-
-export const isCodyEnabled = (): boolean => window.context?.codyEnabled