+++ enterprise/internal/own/search/filter_job_test.go
@@ -320,7 +320,7 @@ func TestApplyCodeOwnershipFiltering(t *testing.T) {
 			}),
 		},
 		{
-			name: "selects results with AND-ed owners specified",
+			name: "selects results with AND-ed include owners specified",
 			args: args{
 				includeOwners: []string{"assigned", "codeowner"},
 				excludeOwners: []string{},
@@ -367,6 +367,194 @@ func TestApplyCodeOwnershipFiltering(t *testing.T) {
 				},
 			}),
 		},
+		{
+			name: "selects results with exclude owner and include owner specified",
+			args: args{
+				includeOwners: []string{"codeowner"},
+				excludeOwners: []string{"assigned"},
+				matches: []result.Match{
+					&result.FileMatch{
+						File: result.File{
+							// assigned owns src/main,
+							// but @codeowner does not own the file
+							Path: "src/main/onlyAssigned.md",
+						},
+					},
+					&result.FileMatch{
+						File: result.File{
+							// @codeowner owns all go files,
+							// and assigned owns src/main
+							Path: "src/main/bothMatch.go",
+						},
+					},
+					&result.FileMatch{
+						File: result.File{
+							// @codeowner owns all go files
+							// but assigned only owns src/main
+							// and this is in src/test.
+							Path: "src/test/onlyCodeowner.go",
+						},
+					},
+				},
+				repoContent: map[string]string{
+					"CODEOWNERS": "*.go @codeowner",
+				},
+			},
+			setup: assignedOwnerSetup(
+				"src/main",
+				&types.User{
+					ID:       42,
+					Username: "assigned",
+				},
+			),
+			want: autogold.Expect([]result.Match{
+				&result.FileMatch{
+					File: result.File{
+						Path: "src/test/onlyCodeowner.go",
+					},
+				},
+			}),
+		},
+		{
+			name: "selects results with AND-ed exclude owners specified",
+			args: args{
+				includeOwners: []string{},
+				excludeOwners: []string{"assigned", "codeowner"},
+				matches: []result.Match{
+					&result.FileMatch{
+						File: result.File{
+							// assigned owns src/main,
+							// but @codeowner does not own the file
+							Path: "src/main/onlyAssigned.md",
+						},
+					},
+					&result.FileMatch{
+						File: result.File{
+							// @codeowner owns all go files,
+							// and assigned owns src/main
+							Path: "src/main/bothMatch.go",
+						},
+					},
+					&result.FileMatch{
+						File: result.File{
+							// @codeowner owns all go files
+							// but assigned only owns src/main
+							// and this is in src/test.
+							Path: "src/test/onlyCodeowner.go",
+						},
+					},
+					&result.FileMatch{
+						File: result.File{
+							// @codeowner owns all go files
+							// but assigned only owns src/main
+							// and this is in src/test.
+							Path: "src/test/noOwners.txt",
+						},
+					},
+				},
+				repoContent: map[string]string{
+					"CODEOWNERS": "*.go @codeowner",
+				},
+			},
+			setup: assignedOwnerSetup(
+				"src/main",
+				&types.User{
+					ID:       42,
+					Username: "assigned",
+				},
+			),
+			want: autogold.Expect([]result.Match{
+				&result.FileMatch{
+					File: result.File{
+						Path: "src/test/noOwners.txt",
+					},
+				},
+			}),
+		},
+		{
+			name: "match commits where any file is owned by included owner",
+			args: args{
+				includeOwners: []string{"@owner"},
+				excludeOwners: []string{},
+				matches: []result.Match{
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file1.notOwned", "file2.owned"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file3.notOwned", "file4.notOwned"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file5.owned"},
+					},
+				},
+				repoContent: map[string]string{
+					"CODEOWNERS": "*.owned @owner\n",
+				},
+			},
+			want: autogold.Expect([]result.Match{
+				&result.CommitMatch{
+					ModifiedFiles: []string{"file1.notOwned", "file2.owned"},
+				},
+				&result.CommitMatch{
+					ModifiedFiles: []string{"file5.owned"},
+				},
+			}),
+		},
+		{
+			name: "discard commits where any file is owned by excluded owner",
+			args: args{
+				includeOwners: []string{},
+				excludeOwners: []string{"@owner"},
+				matches: []result.Match{
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file1.notOwned", "file2.owned"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file3.notOwned", "file4.notOwned"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file5.owned"},
+					},
+				},
+				repoContent: map[string]string{
+					"CODEOWNERS": "*.owned @owner\n",
+				},
+			},
+			want: autogold.Expect([]result.Match{
+				&result.CommitMatch{
+					ModifiedFiles: []string{"file3.notOwned", "file4.notOwned"},
+				},
+			}),
+		},
+		{
+			name: "discard commits through exclude owners despite having include owners",
+			args: args{
+				includeOwners: []string{"@includeOwner"},
+				excludeOwners: []string{"@excludeOwner"},
+				matches: []result.Match{
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file1.included", "file2"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file3.included", "file4.excluded"},
+					},
+					&result.CommitMatch{
+						ModifiedFiles: []string{"file5.excluded", "file3"},
+					},
+				},
+				repoContent: map[string]string{
+					"CODEOWNERS": strings.Join([]string{
+						"*.included @includeOwner",
+						"*.excluded @excludeOwner",
+					}, "\n"),
+				},
+			},
+			want: autogold.Expect([]result.Match{
+				&result.CommitMatch{
+					ModifiedFiles: []string{"file1.included", "file2"},
+				},
+			}),
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -402,6 +590,9 @@ func TestApplyCodeOwnershipFiltering(t *testing.T) {
 			userExternalAccountsStore.ListFunc.SetDefaultReturn(nil, nil)
 			db.UserExternalAccountsFunc.SetDefaultReturn(userExternalAccountsStore)
 			db.TeamsFunc.SetDefaultReturn(database.NewMockTeamStore())
+			repoStore := database.NewMockRepoStore()
+			repoStore.GetFunc.SetDefaultReturn(&types.Repo{ExternalRepo: api.ExternalRepoSpec{ServiceType: "github"}}, nil)
+			db.ReposFunc.SetDefaultReturn(repoStore)
 			if tt.setup != nil {
 				tt.setup(db)
 			}